pyproject.toml
[project]
name = "reminder-bot"
version = "0.1.0"
description = "Your own reminder agent so that you don't miss anything while hanging out on your favorite platform!"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py==2.5.2",
    "pip>=25.1.1",
    "python-dotenv==1.1.0",
    "pytz==2024.1",
    "spacy==3.8.7",
]
